Migrate to Spring Boot.

+++ src/main/java/com/vaddemgen/intelligenthouseclient/bme280/util/Bme280Utils.java
@@ -142,10 +142,10 @@ public final class Bme280Utils {
     }
 
     return Bme280Value.builder()
-        .celsiusTemp(cTemp)
-        .fahrenheitTemp(fTemp)
-        .pressure(pressure)
-        .humidity(humidity)
+        .celsiusTemp((float) cTemp)
+        .fahrenheitTemp((float) fTemp)
+        .pressure((float) pressure)
+        .humidity((float) humidity)
         .build();
   }
 }
